Internal: Tweak Editor menu [ED-22019]
