+++ modules/editor-one/assets/js/sidebar-navigation/components/shared/styled-components.js
@@ -5,7 +5,7 @@ export const NavContainer = styled( Box )( ( { theme } ) => ( {
 	flexDirection: 'column',
 	height: '100%',
 	backgroundColor: theme.palette.background.paper,
-	borderRight: `1px solid ${ theme.palette.divider }`,
+	borderInlineEnd: `1px solid ${ theme.palette.divider }`,
 } ) );
 
 export const SiteIconBox = styled( Box )( ( { theme } ) => ( {
@@ -24,7 +24,7 @@ export const SiteIconBox = styled( Box )( ( { theme } ) => ( {
 
 export const CollapseButton = styled( IconButton )( ( { theme } ) => ( {
 	position: 'absolute',
-	right: -12,
+	insetInlineEnd: -12,
 	bottom: -12,
 	width: 24,
 	height: 24,
@@ -45,4 +45,3 @@ export const ScrollableContent = styled( Box )( {
 	overflowY: 'auto',
 	overflowX: 'hidden',
 } );
-
