+++ modules/editor-one/assets/scss/sidebar-navigation.scss
@@ -1,5 +1,6 @@
 .e-has-sidebar-navigation {
 	--editor-one-sidebar-admin-top-bar-height: 50px;
+	--editor-one-wordpress-admin-bar-height: 32px;
 	--editor-one-sidebar-navigation-width: 240px;
 	--editor-one-sidebar-navigation-width-collapsed: 56px;
 	--editor-one-sidebar-left-offset: 160px;
@@ -7,12 +8,13 @@
 	--editor-one-sidebar-z-index: 99;
 	--editor-one-sidebar-bg: #fff;
 	--editor-one-sidebar-content-padding: 20px;
+	--editor-one-sidebar-transition: width 0.3s ease-in-out, padding-inline-start 0.3s ease-in-out, margin-inline-start 0.3s ease-in-out;
 }
 
 #editor-one-sidebar-navigation {
-	position: absolute;
-	top: var(--editor-one-sidebar-admin-top-bar-height);
-	left: var(--editor-one-sidebar-left-offset);
+	position: fixed;
+	top: calc(var(--editor-one-sidebar-admin-top-bar-height) + var(--editor-one-wordpress-admin-bar-height));
+	inset-inline-start: var(--editor-one-sidebar-left-offset);
 	bottom: 0;
 	width: var(--editor-one-sidebar-navigation-width);
 	z-index: var(--editor-one-sidebar-z-index);
@@ -20,11 +22,11 @@
 	overflow: visible;
 
 	.folded & {
-		left: var(--editor-one-sidebar-left-offset-collapsed);
+		inset-inline-start: var(--editor-one-sidebar-left-offset-collapsed);
 	}
 
 	@media screen and (max-width: 960px) {
-		left: var(--editor-one-sidebar-left-offset-collapsed);
+		inset-inline-start: var(--editor-one-sidebar-left-offset-collapsed);
 	}
 
 	@media screen and (max-width: 782px) {
@@ -36,49 +38,67 @@
 	}
 }
 
+.e-sidebar-transitioning {
+	#editor-one-sidebar-navigation {
+		transition: var(--editor-one-sidebar-transition);
+	}
+
+	#wpcontent,
+	#wpbody-content,
+	#wpfooter {
+		transition: var(--editor-one-sidebar-transition);
+	}
+}
+
 .e-has-sidebar-navigation {
 	#wpcontent {
-		padding-left: var(--editor-one-sidebar-navigation-width);
+		padding-inline-start: var(--editor-one-sidebar-navigation-width);
 	}
 
 	&.folded #wpcontent {
-		padding-left: var(--editor-one-sidebar-navigation-width);
+		padding-inline-start: var(--editor-one-sidebar-navigation-width);
 	}
 
 	&.e-sidebar-is-collapsed {
 		#wpcontent {
-			padding-left: var(--editor-one-sidebar-navigation-width-collapsed);
+			padding-inline-start: var(--editor-one-sidebar-navigation-width-collapsed);
 		}
 
 		&.folded #wpcontent {
-			padding-left: var(--editor-one-sidebar-navigation-width-collapsed);
+			padding-inline-start: var(--editor-one-sidebar-navigation-width-collapsed);
 		}
 
 		#wpbody-content {
-			padding-left: var(--editor-one-sidebar-content-padding);
+			padding-inline-start: var(--editor-one-sidebar-content-padding);
 		}
 
 		#wpfooter {
-			margin-left: var(--editor-one-sidebar-navigation-width-collapsed);
+			margin-inline-start: var(--editor-one-sidebar-navigation-width-collapsed);
 		}
 	}
 
 	@media screen and (max-width: 782px) {
 		#wpcontent {
-			padding-left: 0;
+			padding-inline-start: 0;
 		}
 
 		#wpfooter {
-			margin-left: 0;
+			margin-inline-start: 0;
 		}
 	}
 
 	#wpbody-content {
-		padding-left: var(--editor-one-sidebar-content-padding);
+		padding-inline-start: var(--editor-one-sidebar-content-padding);
 	}
 
 	#wpfooter {
-		margin-left: var(--editor-one-sidebar-navigation-width);
+		margin-inline-start: var(--editor-one-sidebar-navigation-width);
+
+		#footer-left {
+			float: none;
+			text-align: center;
+			margin: 0 auto;
+		}
 	}
 
 	#adminmenu {
